Change total column data type to float in carts table

+++ pc-parts-api/database/migrations/2024_03_28_030010_create_carts_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
     {
         Schema::create('carts', function (Blueprint $table) {
             $table->id();
-            $table->integer('total');
+            $table->float('total');
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
             $table->timestamps();
         });
